[api] Move intrinsic functions into IntrinsicFunction class
